adding accordion to form

+++ opsml/registry/sql/migration/alembic/versions/2d474549026b_.py
@@ -21,17 +21,30 @@ depends_on = None
 
 
 def upgrade() -> None:
+    logger.info(f"Alembic revision: add auditcard_uid and uris cols - {revision}")
+
     bind = op.get_context().bind
     insp = sa.inspect(bind)
     columns = insp.get_columns(RegistryTableNames.DATA.value)
 
-    if not "uris" in [column["name"] for column in columns]:
-        logger.info("Migration Adding uris column to data table")
-        with op.batch_alter_table(RegistryTableNames.DATA.value) as batch_op:
-            batch_op.add_column(sa.Column("uris", sa.JSON))
+    for table_name in [RegistryTableNames.DATA.value, RegistryTableNames.MODEL.value]:
+        columns = insp.get_columns(table_name)
+        if not "auditcard_uid" in [column["name"] for column in columns]:
+            logger.info(f"Migration Adding auditcard column to {table_name} table")
+            with op.batch_alter_table(table_name) as batch_op:
+                batch_op.add_column(sa.Column("auditcard_uid", sa.String(2048)))
+
+                if table_name == RegistryTableNames.DATA.value:
+                    if not "uris" in [column["name"] for column in columns]:
+                        logger.info("Migration Adding uris column to data table")
+                        batch_op.add_column(sa.Column("uris", sa.JSON))
 
 
 def downgrade() -> None:
     logger.info("Dropping uris column from data table")
-    with op.batch_alter_table(RegistryTableNames.DATA.value) as batch_op:
-        batch_op.drop_column("uris")
+    for table_name in [RegistryTableNames.DATA.value, RegistryTableNames.MODEL.value]:
+        with op.batch_alter_table(table_name) as batch_op:
+            batch_op.drop_column("auditcard_uid")
+
+            if table_name == RegistryTableNames.DATA.value:
+                batch_op.drop_column("uris")
